Database Encryption Support
CouchDB can optionally encrypt databases and views.

We use AES in Counter Mode, which ensures we can encrypt and decrypt
any section of the file without padding or alignment. The ciphertext
is the same length as the plaintext. This mode provides
confidentiality but not authentication.

Key management is configurable, a system administrator can write a
module implementing the aegis_key_manager behaviour with any
implementation. CouchDB ships with an implementation that stores keys
in the config file as an example, but this is not suitable for
production.